+++ far/interf.hpp
@@ -143,7 +143,7 @@ void VText(const WCHAR *Str);
 void HiText(const WCHAR *Str,int HiColor,int isVertText=0);
 void vmprintf(const WCHAR *fmt,...);
 void PutText(int X1,int Y1,int X2,int Y2,const void *Src);
-void GetText(int X1,int Y1,int X2,int Y2,void *Dest,int DestSize);
+void GetText(int X1,int Y1,int X2,int Y2,void *Dest,size_t DestSize);
 void BoxText(wchar_t Chr);
 void BoxText(const wchar_t *Str,int IsVert=0);
