+++ far/scrbuf.hpp
@@ -75,7 +75,7 @@ class ScreenBuf
 
 	public:
 		void FillBuf();
-		void Read(int X1,int Y1,int X2,int Y2,CHAR_INFO *Text,int MaxTextLength);
+		void Read(int X1,int Y1,int X2,int Y2,CHAR_INFO *Text,size_t MaxTextLength);
 		void Write(int X,int Y,const CHAR_INFO *Text,int TextLength);
 		void RestoreMacroChar();
 		void RestoreElevationChar();
